Move azure storage resource to bicep model
