+++ src/Aspire.Hosting.Azure/AzureQueueStorageResource.cs
@@ -1,7 +1,10 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Aspire.Hosting.ApplicationModel;
+using Aspire.Hosting.ApplicationModel;
+using Aspire.Hosting.Publishing;
+
+namespace Aspire.Hosting.Azure;
 
 /// <summary>
 /// Represents an Azure Queue Storage resource.
@@ -13,13 +16,29 @@ public class AzureQueueStorageResource(string name, AzureStorageResource storage
     IResourceWithParent<AzureStorageResource>
 {
     /// <summary>
-    /// Gets the parent AzureStorageResource of this AzureQueueStorageResource.
+    /// Gets the parent AzureStorageResource of this AzureBlobStorageResource.
     /// </summary>
     public AzureStorageResource Parent => storage;
 
     /// <summary>
-    /// Gets the connection string for the Azure Queue Storage resource.
+    /// Gets the connection string template for the manifest for the Azure Blob Storage resource.
     /// </summary>
-    /// <returns>The connection string for the Azure Queue Storage resource.</returns>
+    public string ConnectionStringExpression => Parent.QueueEndpoint.ValueExpression;
+
+    /// <summary>
+    /// Gets the connection string for the Azure Blob Storage resource.
+    ///</summary>
+    ///<returns> The connection string for the Azure Blob Storage resource.</returns>
     public string? GetConnectionString() => Parent.GetQueueConnectionString();
+
+    /// <summary>
+    /// TODO: Doc Comments
+    /// </summary>
+    /// <param name="context"></param>
+    public void WriteToManifest(ManifestPublishingContext context)
+    {
+        context.Writer.WriteString("type", "value.v0");
+        context.Writer.WriteString("connectionString", ConnectionStringExpression);
+        context.Writer.WriteString("parent", Parent.Name);
+    }
 }
