+++ src/Aspire.Hosting.Azure/AzureResourceExtensions.cs
@@ -15,160 +15,6 @@ namespace Aspire.Hosting;
 public static class AzureResourceExtensions
 {
     /// <summary>
-    /// Adds an Azure Storage resource to the application model. This resource can be used to create Azure blob, table, and queue resources.
-    /// </summary>
-    /// <param name="builder">The <see cref="IDistributedApplicationBuilder"/>.</param>
-    /// <param name="name">The name of the resource.</param>
-    /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureStorageResource> AddAzureStorage(this IDistributedApplicationBuilder builder, string name)
-    {
-        var resource = new AzureStorageResource(name);
-        return builder.AddResource(resource)
-                      .WithManifestPublishingCallback(WriteAzureStorageToManifest);
-    }
-
-    private static void WriteAzureStorageToManifest(ManifestPublishingContext context)
-    {
-        context.Writer.WriteString("type", "azure.storage.v0");
-    }
-
-    /// <summary>
-    /// Adds an Azure blob storage resource to the application model. This resource requires an <see cref="AzureStorageResource"/> to be added to the application model.
-    /// </summary>
-    /// <param name="storageBuilder">The Azure storage resource builder.</param>
-    /// <param name="name">The name of the resource. This name will be used as the connection string name when referenced in a dependency.</param>
-    /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureBlobStorageResource> AddBlobs(this IResourceBuilder<AzureStorageResource> storageBuilder, string name)
-    {
-        var resource = new AzureBlobStorageResource(name, storageBuilder.Resource);
-        return storageBuilder.ApplicationBuilder.AddResource(resource)
-                             .WithManifestPublishingCallback(context => WriteBlobStorageToManifest(context, resource));
-    }
-
-    private static void WriteBlobStorageToManifest(ManifestPublishingContext context, AzureBlobStorageResource resource)
-    {
-        context.Writer.WriteString("type", "azure.storage.blob.v0");
-        context.Writer.WriteString("parent", resource.Parent.Name);
-    }
-
-    /// <summary>
-    /// Adds an Azure table storage resource to the application model. This resource requires an <see cref="AzureStorageResource"/> to be added to the application model.
-    /// </summary>
-    /// <param name="storageBuilder">The Azure storage resource builder.</param>
-    /// <param name="name">The name of the resource. This name will be used as the connection string name when referenced in a dependency.</param>
-    /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureTableStorageResource> AddTables(this IResourceBuilder<AzureStorageResource> storageBuilder, string name)
-    {
-        var resource = new AzureTableStorageResource(name, storageBuilder.Resource);
-        return storageBuilder.ApplicationBuilder.AddResource(resource)
-                             .WithManifestPublishingCallback(context => WriteTableStorageToManifest(context, resource));
-    }
-
-    private static void WriteTableStorageToManifest(ManifestPublishingContext context, AzureTableStorageResource resource)
-    {
-        context.Writer.WriteString("type", "azure.storage.table.v0");
-        context.Writer.WriteString("parent", resource.Parent.Name);
-    }
-
-    /// <summary>
-    /// Adds an Azure queue storage resource to the application model. This resource requires an <see cref="AzureStorageResource"/> to be added to the application model.
-    /// </summary>
-    /// <param name="builder">The Azure storage resource builder.</param>
-    /// <param name="name">The name of the resource. This name will be used as the connection string name when referenced in a dependency.</param>
-    /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureQueueStorageResource> AddQueues(this IResourceBuilder<AzureStorageResource> builder, string name)
-    {
-        var resource = new AzureQueueStorageResource(name, builder.Resource);
-        return builder.ApplicationBuilder.AddResource(resource)
-                             .WithManifestPublishingCallback(context => WriteQueueStorageToManifest(context, resource));
-    }
-
-    private static void WriteQueueStorageToManifest(ManifestPublishingContext context, AzureQueueStorageResource resource)
-    {
-        context.Writer.WriteString("type", "azure.storage.queue.v0");
-        context.Writer.WriteString("parent", resource.Parent.Name);
-    }
-
-    /// <summary>
-    /// Configures an Azure Storage resource to be emulated using Azurite. This resource requires an <see cref="AzureStorageResource"/> to be added to the application model. This version the package defaults to version 3.29.0 of the mcr.microsoft.com/azure-storage/azurite container image.
-    /// </summary>
-    /// <param name="builder">The Azure storage resource builder.</param>
-    /// <param name="configureContainer">Callback that exposes underlying container used for emulation to allow for customization.</param>
-    /// <returns>A reference to the <see cref="IResourceBuilder{T}"/>.</returns>
-    public static IResourceBuilder<AzureStorageResource> UseEmulator(this IResourceBuilder<AzureStorageResource> builder, Action<IResourceBuilder<AzureStorageEmulatorResource>>? configureContainer = null)
-    {
-        builder.WithAnnotation(new EndpointAnnotation(ProtocolType.Tcp, name: "blob", containerPort: 10000))
-               .WithAnnotation(new EndpointAnnotation(ProtocolType.Tcp, name: "queue", containerPort: 10001))
-               .WithAnnotation(new EndpointAnnotation(ProtocolType.Tcp, name: "table", containerPort: 10002))
-               .WithAnnotation(new ContainerImageAnnotation { Image = "mcr.microsoft.com/azure-storage/azurite", Tag = "3.29.0" });
-
-        if (configureContainer != null)
-        {
-            var surrogate = new AzureStorageEmulatorResource(builder.Resource);
-            var surrogateBuilder = builder.ApplicationBuilder.CreateResourceBuilder(surrogate);
-            configureContainer(surrogateBuilder);
-        }
-
-        return builder;
-    }
-
-    /// <summary>
-    /// Enables persistence in the Azure Storage emulator.
-    /// </summary>
-    /// <param name="builder">The builder for the <see cref="AzureStorageEmulatorResource"/>.</param>
-    /// <param name="path">Relative path to the AppHost where emulator storage is persisted between runs.</param>
-    /// <returns>A builder for the <see cref="AzureStorageEmulatorResource"/>.</returns>
-    public static IResourceBuilder<AzureStorageEmulatorResource> UsePersistence(this IResourceBuilder<AzureStorageEmulatorResource> builder, string? path = null)
-    {
-        path = path ?? $".azurite/{builder.Resource.Name}";
-        var fullyQualifiedPath = Path.GetFullPath(path, builder.ApplicationBuilder.AppHostDirectory);
-        return builder.WithBindMount(fullyQualifiedPath, "/data", isReadOnly: false);
-
-    }
-
-    /// <summary>
-    /// Modifies the host port that the storage emulator listens on for blob requests.
-    /// </summary>
-    /// <param name="builder">Storage emulator resource builder.</param>
-    /// <param name="port">Host port to use.</param>
-    /// <returns></returns>
-    public static IResourceBuilder<AzureStorageEmulatorResource> UseBlobPort(this IResourceBuilder<AzureStorageEmulatorResource> builder, int port)
-    {
-        return builder.WithEndpoint("blob", endpoint =>
-        {
-            endpoint.Port = port;
-        });
-    }
-
-    /// <summary>
-    /// Modifies the host port that the storage emulator listens on for queue requests.
-    /// </summary>
-    /// <param name="builder">Storage emulator resource builder.</param>
-    /// <param name="port">Host port to use.</param>
-    /// <returns></returns>
-    public static IResourceBuilder<AzureStorageEmulatorResource> UseQueuePort(this IResourceBuilder<AzureStorageEmulatorResource> builder, int port)
-    {
-        return builder.WithEndpoint("queue", endpoint =>
-        {
-            endpoint.Port = port;
-        });
-    }
-
-    /// <summary>
-    /// Modifies the host port that the storage emulator listens on for table requests.
-    /// </summary>
-    /// <param name="builder">Storage emulator resource builder.</param>
-    /// <param name="port">Host port to use.</param>
-    /// <returns></returns>
-    public static IResourceBuilder<AzureStorageEmulatorResource> UseTablePort(this IResourceBuilder<AzureStorageEmulatorResource> builder, int port)
-    {
-        return builder.WithEndpoint("table", endpoint =>
-        {
-            endpoint.Port = port;
-        });
-    }
-
-    /// <summary>
     /// Configures an Azure Cosmos DB resource to be emulated using the Azure Cosmos DB emulator with the NoSQL API. This resource requires an <see cref="AzureCosmosDBResource"/> to be added to the application model.
     /// For more information on the Azure Cosmos DB emulator, see <a href="https://learn.microsoft.com/azure/cosmos-db/emulator#authentication"></a>
     /// </summary>
